+++ src/OTPublisher.js
@@ -67,6 +67,8 @@ class OTPublisher extends Component {
         const value = useDefault(this.props.properties[key], defaultValue);
         if (key === 'cameraPosition') {
           OT.changeCameraPosition(this.state.publisherId, value);
+        } else if (key === 'videoContentHint') {
+          OT.changeVideoContentHint(this.state.publisherId, value);
         } else {
           OT[key](this.state.publisherId, value);
         }
@@ -76,6 +78,7 @@ class OTPublisher extends Component {
     updatePublisherProperty('publishAudio', true);
     updatePublisherProperty('publishVideo', true);
     updatePublisherProperty('cameraPosition', 'front');
+    updatePublisherProperty('videoContentHint', '');
   }
   componentWillUnmount() {
     OT.destroyPublisher(this.state.publisherId, (error) => {