+++ src/helpers/OTPublisherHelper.js
@@ -33,6 +33,19 @@ const sanitizeVideoSource = (videoSource = 'camera') => (videoSource === 'camera
 const sanitizeAudioBitrate = (audioBitrate = 40000) =>
   (audioBitrate < 6000 || audioBitrate > 510000 ? 40000 : audioBitrate);
 
+const sanitizeVideoContentHint = (sanitizeVideoContentHint = '') => {
+  switch (sanitizeVideoContentHint) {
+    case 'motion':
+      return 'motion';
+    case 'detail':
+      return 'detail';
+    case 'text':
+      return 'text';
+    default:
+      return '';
+  }
+};
+
 const sanitizeProperties = (properties) => {
   if (typeof properties !== 'object') {
     return {
@@ -44,8 +57,10 @@ const sanitizeProperties = (properties) => {
       cameraPosition: 'front',
       audioFallbackEnabled: true,
       audioBitrate: 40000,
+      enableDtx: false,
       frameRate: 30,
       resolution: sanitizeResolution(),
+      videoContentHint: '',
       videoSource: 'camera',
     };
   }
@@ -58,8 +73,10 @@ const sanitizeProperties = (properties) => {
     cameraPosition: sanitizeCameraPosition(properties.cameraPosition),
     audioFallbackEnabled: sanitizeBooleanProperty(properties.audioFallbackEnabled),
     audioBitrate: sanitizeAudioBitrate(properties.audioBitrate),
+    enableDtx: sanitizeBooleanProperty(properties.enableDtx ? properties.enableDtx : false),
     frameRate: sanitizeFrameRate(properties.frameRate),
     resolution: sanitizeResolution(properties.resolution),
+    videoContentHint: sanitizeVideoContentHint(properties.videoContentHint),
     videoSource: sanitizeVideoSource(properties.videoSource),
   };
 };